requirements.txt
# Vietnam Stock Pipeline - Big Data Requirements
# Senior Data Engineer Level - Kafka + Spark + VNStock

# Core dependencies - Pinned versions for security
streamlit==1.28.1
pandas==2.0.3
numpy==1.24.3
sqlalchemy==2.0.19
psycopg2-binary==2.9.7
python-dotenv==1.0.0

# VNStock - Vietnam Stock Market Data Library
vnstock==0.2.9.2.3
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
ipython==8.12.0

# Kafka - Message Streaming
kafka-python==2.0.2

# Spark - Big Data Processing
pyspark==3.5.0

# Data visualization
plotly==5.15.0

# Utilities
pytz==2023.3
structlog==23.1.0
psutil==5.9.5
schedule==1.1.0